requirements-dev.txt
aiofiles==0.8.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.3.4
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==24.2.0
Babel==2.15.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
colorlog==6.6.0
comm==0.2.2
common==0.1.2
crypto==1.4.1
cryptography==43.0.1
DateTime==5.5
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
DingtalkChatbot==1.5.3
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
execnet==1.9.0
executing==2.0.1
Faker==9.8.3
fastjsonschema==2.20.0
Flask==1.1.4
fqdn==1.5.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.4.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.3
importlib_metadata==8.0.0
iniconfig==1.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
isoduration==20.11.0
itchat==1.3.10
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpath==0.82
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaitaistruct==0.9
ldap3==2.8.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
mistune==3.0.2
mitmproxy==5.3.0
msgpack==1.0.3
multidict==6.0.2
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
openpyxl==3.0.9
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
paramiko==3.4.1
parso==0.8.4
passlib==1.7.4
pefile==2023.2.7
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.19.4
psutil==6.0.0
psycopg2==2.9.9
publicsuffix2==2.20191221
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pydantic_core==2.20.1
pyDes==2.0.1
pydivert==2.1.0
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypng==0.0.21
PyQRCode==1.2.1
pytest==7.1.3
pytest-forked==1.3.0
pytest-xdist==2.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==304
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
redis==4.3.6
referencing==0.35.1
requests==2.26.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sanic==22.3.1
sanic-routing==22.3.0
Send2Trash==1.8.3
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==1.26.7
urwid==2.1.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==4.0.11
wsproto==1.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwings==0.27.7
xlwt==1.3.0
yarg==0.1.9
zipp==3.19.2
zope.interface==7.0.3
zstandard==0.17.0
